# Optimize packet handling in client and server by caching type to processor lookup