Performance difference can be seen here, where `.MakeGenericType(...)` lookups take ~0.2ms and cached lookups ~0.-0005 to ~0.001ms. A **~400x speedup** but only for the lookup part, not the processing itself.

This should reduce latency issues.

```log
[11:36:57.897] [DBG] Perf: Processor lookup took 0.2477ms
[11:37:38.220] [DBG] Perf: Processor lookup took 0.3134ms
[11:37:38.221] [DBG] Perf: Processor lookup took 0.2795ms
[11:37:38.237] [DBG] Perf: Processor lookup took 0.0077ms
[11:37:38.237] [DBG] Perf: Processor lookup took 0.0003ms
[11:37:38.246] [DBG] Perf: Processor lookup took 0.0005ms
[11:37:38.246] [DBG] Perf: Processor lookup took 0.0003ms
[11:37:38.295] [DBG] Perf: Processor lookup took 0.0006ms
[11:37:38.338] [DBG] Perf: Processor lookup took 0.4232ms
[11:37:38.812] [DBG] Perf: Processor lookup took 0.0007ms
```